rework brightness steps
This sets up a cascade of brightness calibration sliders. The
first one finds a solid color for 1/2 of full scale brightness by
matching a 50% dither of 0x000000 (black) with 0xFFFFFF (white).
The next finds a value for 1/4 brightness (50% dither of 1/2
brightness value with black. The rest continue in that pattern.